+++ packages/get-file/src/downloadFile.ts
@@ -6,21 +6,23 @@
  * @copyright 2019-2021 Nguyen Huu Phuoc <me@phuoc.ng>
  */
 
-import type { OpenFile } from '@react-pdf-viewer/core';
+import type { OpenFile, PdfJs } from '@react-pdf-viewer/core';
 
-export const downloadFile = (file: OpenFile, saveAs: string): void => {
-    const blobUrl =
-        typeof file.data === 'string' ? '' : URL.createObjectURL(new Blob([file.data], { type: 'application/pdf' }));
-    const link = document.createElement('a');
-    link.style.display = 'none';
-    link.href = blobUrl || file.name;
-    link.setAttribute('download', saveAs);
+export const downloadFile = (doc: PdfJs.PdfDocument, file: OpenFile, saveAs: string): void => {
+    doc.getData().then((data) => {
+        const blobUrl =
+            typeof file.data === 'string' ? '' : URL.createObjectURL(new Blob([data], { type: 'application/pdf' }));
+        const link = document.createElement('a');
+        link.style.display = 'none';
+        link.href = blobUrl || file.name;
+        link.setAttribute('download', saveAs);
 
-    document.body.appendChild(link);
-    link.click();
+        document.body.appendChild(link);
+        link.click();
 
-    document.body.removeChild(link);
-    if (blobUrl) {
-        URL.revokeObjectURL(blobUrl);
-    }
+        document.body.removeChild(link);
+        if (blobUrl) {
+            URL.revokeObjectURL(blobUrl);
+        }
+    });
 };