+++ CHANGELOG.md
@@ -19,6 +19,7 @@ const defaultLayoutPluginInstance = defaultLayoutPlugin({
 **Bug fixes**
 
 -   Can't use special characters such as `(`, `)`, `[`, `]`, `*` in the keyword
+-   The Get File plugin doesn't work with a file which is opened with credentials
 -   The highlight area aren't displayed
 -   The `selectedText` prop of `RenderHighlightContentProps` isn't correct